[HW Node 04-aut][fix] - added
auth.js
owner in contacts

+++ models/contacts.js
@@ -1,44 +1,45 @@
 const Contact = require("../service/schemas/contact");
 
-const listContacts = async () => {
+const listContacts = async (owner) => {
   try {
-    return await Contact.find();
+    return await Contact.find({ owner });
   } catch (error) {
     console.log("Error getting contact list: ", error);
     throw error;
   }
 };
 
-const getContactById = async (contactId) => {
+const getContactById = async (contactId, owner) => {
   try {
-    return await Contact.findOne({ _id: contactId });
+    return await Contact.findOne({ _id: contactId, owner });
   } catch (error) {
     console.log(`Error getting contact with id ${contactId}: `, error);
     throw error;
   }
 };
 
-const removeContact = async (contactId) => {
+const removeContact = async (contactId, owner) => {
   try {
-    return await Contact.findByIdAndRemove({ _id: contactId });
+    return await Contact.findByIdAndRemove({ _id: contactId, owner });
   } catch (error) {
     console.log(`Error removing contact with id ${contactId}: `, error);
     throw error;
   }
 };
 
-const addContact = async (body) => {
+const addContact = async (body, owner) => {
   try {
-    return await Contact.create(body);
+    const contact = { body, owner };
+    return await Contact.create(contact);
   } catch (error) {
     console.log("Error adding new contact: ", error);
     throw error;
   }
 };
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (contactId, body, owner) => {
   try {
-    return await Contact.findByIdAndUpdate({ _id: contactId }, body, {
+    return await Contact.findByIdAndUpdate({ _id: contactId, owner }, body, {
       new: true,
     });
   } catch (error) {
@@ -47,10 +48,10 @@ const updateContact = async (contactId, body) => {
   }
 };
 
-const updatedStatusContact = async (contactId, favorite) => {
+const updatedStatusContact = async (contactId, favorite, owner) => {
   try {
     return await Contact.findByIdAndUpdate(
-      { _id: contactId },
+      { _id: contactId, owner },
       { $set: { favorite: favorite } },
       { new: true }
     );
